Spring API meeting 2 progress
https: //github.com/acmcsufoss/i/issues/18
